trunk, idl:
Summary: Added the the initial version of the native widget API.

The maWidget* syscalls is an interface for creating and manipulating native widgets. The maWidgetCreate syscall creates a native widget without showing it on the screen, while maWidgetAddChild actually adds the widget to a view tree and displays it. The maWidgetPropertySet syscall handles the manipulation of native widgets, like setting the size and location. But it can also be used to set the content of a widget.

Widgets can notify the runtime when it's state changes by sending a MA_EVENT_TYPE_WIDGET, which data field points to a MAWidgetEventData struct. The MAWidgetEventData struct further specifies the event type and identifies the widget that sent the event.

Note: This API is likely to change.


git-svn-id: svn://localhost/public/MoSync/branches/new-native-ui@2445 6ad7a17e-d552-4a1e-ace4-f958fe14b117