Summary: Updated the newlib crtlib.s
The crtlib.s that is used with newlib has diverged from the one in MAStd, which means that the code for multiple entry points does not work. This was discovered when running testify with newlib, causing a 'Multiple entry points" error message to occur. Fixed by updating the newlib crtlib.s.