Commented and shaped up the code.
Status:
Pure syscall system works.
Widget system partially works. Preview works fine with one camera, but has some problems correctly switching the cameras on the same widget.
Properties, snapshot, and video capture not implemented.
Needs to correctly stop and start the cameras when the app goes to the background, or is terminated.